+++ cmd/mq_json/main.go
@@ -30,7 +30,7 @@ var BuildStamp string
 var GitCommit string
 
 func initLog() {
-	level, err := log.ParseLevel(config.logLevel)
+	level, err := log.ParseLevel(config.cf.LogLevel)
 	if err != nil {
 		level = log.InfoLevel
 	}
@@ -65,7 +65,7 @@ func main() {
 		log.Errorln("Must provide a queue manager name to connect to.")
 		os.Exit(1)
 	}
-	d, err := time.ParseDuration(config.interval + "s")
+	d, err := time.ParseDuration(config.interval)
 	if err != nil {
 		log.Errorln("Invalid value for interval parameter: ", err)
 		os.Exit(1)