+++ cmd/mq_prometheus/exporter.go
@@ -122,7 +122,7 @@ func (e *exporter) Collect(ch chan<- prometheus.Metric) {
 	pollStatus := false
 	thisPoll := time.Now()
 	elapsed := thisPoll.Sub(lastPoll)
-	if elapsed >= config.pollIntervalDuration {
+	if elapsed >= config.cf.PollIntervalDuration {
 		log.Debugf("Polling for object status")
 		lastPoll = thisPoll
 		pollStatus = true