+++ pkg/config/config.go
@@ -25,6 +25,7 @@ import (
 	"flag"
 	"fmt"
 	"github.com/ibm-messaging/mq-golang/mqmetric"
+	"time"
 )
 
 // Configuration attributes shared by all the monitor sample programs
@@ -33,6 +34,8 @@ type Config struct {
 	ReplyQ   string
 
 	MetaPrefix string
+
+	TZOffsetString string
 
 	MonitoredQueues            string
 	MonitoredQueuesFile        string
@@ -44,11 +47,23 @@ type Config struct {
 	MonitoredSubscriptionsFile string
 
 	UseStatus bool
+	LogLevel  string
+
+	// This is used for DISPLAY xxSTATUS commands, but not the collection of published resource stats
+	pollInterval         string
+	PollIntervalDuration time.Duration
 
 	CC mqmetric.ConnectionConfig
 }
 
+const (
+	defaultPollInterval = "0s"
+	defaultTZOffset     = "0h"
+)
+
 func InitConfig(cm *Config) {
+	flag.StringVar(&cm.LogLevel, "log.level", "error", "Log level - debug, info, error")
+
 	flag.StringVar(&cm.QMgrName, "ibmmq.queueManager", "", "Queue Manager name")
 	flag.StringVar(&cm.ReplyQ, "ibmmq.replyQueue", "SYSTEM.DEFAULT.MODEL.QUEUE", "Reply Queue to collect data")
 
@@ -64,13 +79,17 @@ func InitConfig(cm *Config) {
 	flag.StringVar(&cm.MonitoredSubscriptions, "ibmmq.monitoredSubscriptions", "*", "Patterns of subscriptions to monitor")
 	flag.StringVar(&cm.MonitoredSubscriptionsFile, "ibmmq.monitoredSubscriptionsFile", "", "File with patterns of subscriptions to monitor")
 
+	// qStatus was the original flag but prefer to use useStatus as more meaningful for all object types
 	flag.BoolVar(&cm.UseStatus, "ibmmq.qStatus", false, "Add metrics from the QSTATUS fields")
-	flag.BoolVar(&cm.UseStatus, "ibmmq.useStatus", false, "Add metrics from the object STATUS fields")
+	flag.BoolVar(&cm.UseStatus, "ibmmq.useStatus", false, "Add metrics from all object STATUS fields")
 
 	flag.StringVar(&cm.CC.UserId, "ibmmq.userid", "", "UserId for MQ connection")
 	// If password is not given on command line (and it shouldn't be) then there's a prompt for stdin
 	flag.StringVar(&cm.CC.Password, "ibmmq.password", "", "Password for MQ connection")
 	flag.BoolVar(&cm.CC.ClientMode, "ibmmq.client", false, "Connect as MQ client")
+
+	flag.StringVar(&cm.TZOffsetString, "ibmmq.tzOffset", defaultTZOffset, "Time difference between collector and queue manager")
+	flag.StringVar(&cm.pollInterval, "pollInterval", defaultPollInterval, "Frequency of issuing object status checks")
 
 }
 
@@ -127,6 +146,21 @@ func VerifyConfig(cm *Config) error {
 	}
 
 	// Do not use VerifyPatterns for monitoredTopics or Subs as they follow a very different style
+	if err == nil {
+		offset, err := time.ParseDuration(cm.TZOffsetString)
+		if err != nil {
+			err = fmt.Errorf("Invalid value for interval parameter: %v", err)
+		} else {
+			cm.CC.TZOffsetSecs = offset.Seconds()
+		}
+	}
+
+	if err == nil {
+		cm.PollIntervalDuration, err = time.ParseDuration(cm.pollInterval)
+		if err != nil {
+			err = fmt.Errorf("Invalid value for interval parameter: %v", err)
+		}
+	}
 
 	return err
 }